Add `exclude` to documentation
I believe this addition to the docs could be useful. Perhaps we should caution about how removal of one module affects the other, I'd be happy to write this up (probably not in readme to keep it less verbose?) if you could outline which default unimodules depend on each other / what interfaces.